refactor: add device token via json over http
Post device token data instead of graphql mutation
